+++ CHANGELOG.md
@@ -9,7 +9,7 @@ _general description_
 - **added** - string utilities (lowercase, split)
 - **changed** - tests are now in standalone files in the tests/ directory and can be run using arguments "-t test1,test2"
 - **added** - coroutine implementation (future, promise, awaitables)
-- **added** - job system
+- **added** - job system with coroutines, parent-children jobs and multithread support
 - **added** - temporary dependency `jthread` while compilers gain supports
 - **added** - thread-safe atomic queue